// UserInterface/my-app/src/Models/ProductionOrder.ts
export interface ProductionOrder {
  Id: number;
  CDate?: string;
  ORDID?: number;
  CUSRID?: number;
  ItemId?: number;
  Date?: string;
  FicheID?: number;
  IsConfirmed?: number;
  MachineID?: number;
  Amount?: number;
  IsCompleted?: number;
  UUSRID?: number;
  UDate?: string;
  UnitID?: number;
  CAID?: number;
  IName?:string;
  UName?:string;
  MCode?: string;
  OrderAmount?: number;
  OCode?:string;
  CAName?: string;
}